# By Vitaliy Borodovsky (21) and others
# Via Vitaliy Borodovsky (9) and others
* commit 'd93cdba800bbad14861daea9c705b89dc272e937 [formerly 1483194a898a12e15b7de72c76ed587420728484]': (32 commits)
  DF-6370 fix landscape cursor draw
  DF-6370 fix landscape creation crash
  DF-6370 update landscape metricies
  DF-6370 fix compilation after merge development
  [DF-6370] Fixes according to review
  [DF-6370] Fixes according to review
  [DF-6370] Fixes according to review
  [DF-6370] Fixes according to review
  [DF-6370] Fixes according to review
  [DF-6370] Fixes according to review
  [DF-6370] Fixes according to review
  [DF-6370] Fixes according to review
  [DF-6370] Polishing. Ability to change quality on the fly
  [DF-6370] Polishing. Ability to change quality on the fly
  [DF-6370] Polishing. Ability to change quality on the fly
  [DF-6370] Polishing. Ability to change quality on the fly
  [DF-6370] Added ability to reload quality when it changed. Do not have overhead in vertices.
  [DF-6370] Fix of landscape export problem
  [DF-6370] Fixes according to review
  [DF-6370] Fixes according to review
  ...


Former-commit-id: 9d627da8f5bd1e11888f2fc220c6c39e097a4dd2

+++ Sources/Internal/Render/Material/NMaterial.h
@@ -334,7 +334,7 @@ public:
      \param[in] indices array of indices to render.
      \param[in] indexCount number of indices to render.
 	 */
-	void Draw(RenderDataObject* renderData, uint16* indices = NULL, uint16 indexCount = 0);
+	void Draw(RenderDataObject* renderData, uint16* indices = NULL, int32 indexCount = 0);
 	
     /**
 	 \brief Sets flag to the material affectively altering its shaders by define
